docs: update icon docs
