+++ src/components/icon/__docs__/docs.mdx
@@ -1,0 +1,166 @@
+import { Meta, Title, IconGallery, IconItem } from '@storybook/addon-docs/';
+import { Icon as IconExample } from '../icon';
+
+<Meta title="Elements/Icon" />
+
+# Icon
+
+## Generic Icons
+
+<IconGallery>
+  <IconItem name="Archive">
+    <IconExample name="Archive" />
+  </IconItem>
+  <IconItem name="ArchiveFilled">
+    <IconExample name="ArchiveFilled" />
+  </IconItem>
+  <IconItem name="Assign">
+    <IconExample name="Assign" />
+  </IconItem>
+  <IconItem name="AssignFilled">
+    <IconExample name="AssignFilled" />
+  </IconItem>
+  <IconItem name="Calendar">
+    <IconExample name="Calendar" />
+  </IconItem>
+  <IconItem name="CalendarFilled">
+    <IconExample name="CalendarFilled" />
+  </IconItem>
+  <IconItem name="CheckmarkBox">
+    <IconExample name="CheckmarkBox" />
+  </IconItem>
+  <IconItem name="CheckmarkCircle">
+    <IconExample name="CheckmarkCircle" />
+  </IconItem>
+  <IconItem name="CheckmarkCircleEmpty">
+    <IconExample name="CheckmarkCircleEmpty" />
+  </IconItem>
+  <IconItem name="Checkmark">
+    <IconExample name="Checkmark" />
+  </IconItem>
+  <IconItem name="ChevronDown">
+    <IconExample name="ChevronDown" />
+  </IconItem>
+  <IconItem name="ChevronLeft">
+    <IconExample name="ChevronLeft" />
+  </IconItem>
+  <IconItem name="ChevronRight">
+    <IconExample name="ChevronRight" />
+  </IconItem>
+  <IconItem name="ChevronUp">
+    <IconExample name="ChevronUp" />
+  </IconItem>
+  <IconItem name="Close">
+    <IconExample name="Close" />
+  </IconItem>
+  <IconItem name="Copy">
+    <IconExample name="Copy" />
+  </IconItem>
+  <IconItem name="Edit">
+    <IconExample name="Edit" />
+  </IconItem>
+  <IconItem name="EditFilled">
+    <IconExample name="EditFilled" />
+  </IconItem>
+  <IconItem name="Export">
+    <IconExample name="Export" />
+  </IconItem>
+  <IconItem name="ExternalLink">
+    <IconExample name="ExternalLink" />
+  </IconItem>
+  <IconItem name="Gear">
+    <IconExample name="Gear" />
+  </IconItem>
+  <IconItem name="GearFilled">
+    <IconExample name="GearFilled" />
+  </IconItem>
+  <IconItem name="Import">
+    <IconExample name="Import" />
+  </IconItem>
+  <IconItem name="Info">
+    <IconExample name="Info" />
+  </IconItem>
+  <IconItem name="InfoFilled">
+    <IconExample name="InfoFilled" />
+  </IconItem>
+  <IconItem name="Minus">
+    <IconExample name="Minus" />
+  </IconItem>
+  <IconItem name="EllipsisHorizontal">
+    <IconExample name="EllipsisHorizontal" />
+  </IconItem>
+  <IconItem name="EllipsisVertical">
+    <IconExample name="EllipsisVertical" />
+  </IconItem>
+  <IconItem name="Plus">
+    <IconExample name="Plus" />
+  </IconItem>
+  <IconItem name="NewContact">
+    <IconExample name="NewContact" />
+  </IconItem>
+  <IconItem name="Participant">
+    <IconExample name="Participant" />
+  </IconItem>
+  <IconItem name="ParticipantFilled">
+    <IconExample name="ParticipantFilled" />
+  </IconItem>
+  <IconItem name="PlusCircle">
+    <IconExample name="PlusCircle" />
+  </IconItem>
+  <IconItem name="Preferences">
+    <IconExample name="Preferences" />
+  </IconItem>
+  <IconItem name="Search">
+    <IconExample name="Search" />
+  </IconItem>
+  <IconItem name="Star">
+    <IconExample name="Star" />
+  </IconItem>
+  <IconItem name="StarFilled">
+    <IconExample name="StarFilled" />
+  </IconItem>
+  <IconItem name="Trash">
+    <IconExample name="Trash" />
+  </IconItem>
+  <IconItem name="TrashFilled">
+    <IconExample name="TrashFilled" />
+  </IconItem>
+</IconGallery>
+
+## Attachment Icons
+
+<IconGallery>
+  <IconItem name="AttachmentArchive">
+    <IconExample name="AttachmentArchive" />
+  </IconItem>
+  <IconItem name="AttachmentCalendar">
+    <IconExample name="AttachmentCalendar" />
+  </IconItem>
+  <IconItem name="AttachmentCode">
+    <IconExample name="AttachmentCode" />
+  </IconItem>
+  <IconItem name="AttachmentExcel">
+    <IconExample name="AttachmentExcel" />
+  </IconItem>
+  <IconItem name="AttachmentGeneric">
+    <IconExample name="AttachmentGeneric" />
+  </IconItem>
+  <IconItem name="AttachmentImage">
+    <IconExample name="AttachmentImage" />
+  </IconItem>
+  <IconItem name="AttachmentMusic">
+    <IconExample name="AttachmentMusic" />
+  </IconItem>
+  <IconItem name="AttachmentPdf">
+    <IconExample name="AttachmentPdf" />
+  </IconItem>
+  <IconItem name="AttachmentPowerpoint">
+    <IconExample name="AttachmentPowerpoint" />
+  </IconItem>
+  <IconItem name="AttachmentVideo">
+    <IconExample name="AttachmentVideo" />
+  </IconItem>
+  <IconItem name="AttachmentWord">
+    <IconExample name="AttachmentWord" />
+  </IconItem>
+</IconGallery>
